Revert "fix: Revert "feat: add wasm memory benchmark to bench.json" (#12204)"
This reverts commit fd323cc9d3f2a1e81f3ea2bc82e36c8fcf02db93.
