+++ barretenberg/.gitignore
@@ -7,3 +7,4 @@ ts/dest
 .idea
 cmake-build-debug
 *_opt.pil
+bench-out
