+++ barretenberg/bootstrap.sh
@@ -8,3 +8,20 @@ source $(git rev-parse --show-toplevel)/ci3/source
 ./cpp/bootstrap.sh $@
 ./ts/bootstrap.sh $@
 ./acir_tests/bootstrap.sh $@
+
+cmd=${1:-}
+if [ $cmd == "bench" ]; then
+    rm -rf bench-out && mkdir -p bench-out
+    ./scripts/combine_benchmarks.py \
+    native ./cpp/bench-out/client_ivc_17_in_20_release.json \
+    native ./cpp/bench-out/client_ivc_release.json \
+    native ./cpp/bench-out/ultra_honk_release.json \
+    wasm ./cpp/bench-out/client_ivc_wasm.json \
+    wasm ./cpp/bench-out/ultra_honk_wasm.json \
+    "" ./cpp/bench-out/client_ivc_op_count.json \
+    "" ./cpp/bench-out/client_ivc_op_count_time.json \
+    wasm ./acir_tests/bench-out/ultra_honk_rec_wasm_memory.txt \
+    > ./bench-out/bb-bench.json
+
+  cache_upload barretenberg-bench-results-$COMMIT_HASH.tar.gz ./bench-out/bb-bench.json
+fi
